+++ lib/rest/bulkexports/v1/export.d.ts
@@ -33,7 +33,7 @@ interface ExportListInstance {
   /**
    * Constructs a export
    *
-   * @param resourceType - The type of communication – Messages, Calls
+   * @param resourceType - The type of communication – Messages, Calls, Conferences, and Participants
    */
   get(resourceType: string): ExportContext;
   jobs?: JobListInstance;
@@ -64,7 +64,7 @@ declare class ExportContext {
    * Use them with caution.
    *
    * @param version - Version of the resource
-   * @param resourceType - The type of communication – Messages, Calls
+   * @param resourceType - The type of communication – Messages, Calls, Conferences, and Participants
    */
   constructor(version: V1, resourceType: string);
 
@@ -92,7 +92,7 @@ declare class ExportInstance extends SerializableClass {
    *
    * @param version - Version of the resource
    * @param payload - The instance payload
-   * @param resourceType - The type of communication – Messages, Calls
+   * @param resourceType - The type of communication – Messages, Calls, Conferences, and Participants
    */
   constructor(version: V1, payload: ExportPayload, resourceType: string);