+++ lib/rest/insights/v1/call/summary.d.ts
@@ -19,10 +19,6 @@ type CallSummaryProcessingState = 'complete'|'partial';
 /**
  * Initialize the CallSummaryList
  *
- * PLEASE NOTE that this class contains preview products that are subject to
- * change. Use them with caution. If you currently do not have developer preview
- * access, please contact help@twilio.com.
- *
  * @param version - Version of the resource
  * @param callSid - The call_sid
  */
@@ -64,6 +60,7 @@ interface CallSummaryResource {
   carrier_edge: object;
   client_edge: object;
   connect_duration: number;
+  created_time: Date;
   duration: number;
   end_time: Date;
   from: object;
@@ -86,10 +83,6 @@ interface CallSummarySolution {
 declare class CallSummaryContext {
   /**
    * Initialize the CallSummaryContext
-   *
-   * PLEASE NOTE that this class contains preview products that are subject to
-   * change. Use them with caution. If you currently do not have developer preview
-   * access, please contact help@twilio.com.
    *
    * @param version - Version of the resource
    * @param callSid - The call_sid
@@ -120,10 +113,6 @@ declare class CallSummaryInstance extends SerializableClass {
   /**
    * Initialize the CallSummaryContext
    *
-   * PLEASE NOTE that this class contains preview products that are subject to
-   * change. Use them with caution. If you currently do not have developer preview
-   * access, please contact help@twilio.com.
-   *
    * @param version - Version of the resource
    * @param payload - The instance payload
    * @param callSid - The call_sid
@@ -139,6 +128,7 @@ declare class CallSummaryInstance extends SerializableClass {
   carrierEdge: any;
   clientEdge: any;
   connectDuration: number;
+  createdTime: Date;
   duration: number;
   endTime: Date;
   /**
@@ -174,10 +164,6 @@ declare class CallSummaryInstance extends SerializableClass {
 declare class CallSummaryPage extends Page<V1, CallSummaryPayload, CallSummaryResource, CallSummaryInstance> {
   /**
    * Initialize the CallSummaryPage
-   *
-   * PLEASE NOTE that this class contains preview products that are subject to
-   * change. Use them with caution. If you currently do not have developer preview
-   * access, please contact help@twilio.com.
    *
    * @param version - Version of the resource
    * @param response - Response from the API