+++ lib/rest/preview/bulk_exports/export.d.ts
@@ -34,7 +34,7 @@ interface ExportListInstance {
   /**
    * Constructs a export
    *
-   * @param resourceType - The type of communication – Messages, Calls
+   * @param resourceType - The type of communication – Messages, Calls, Conferences, and Participants
    */
   get(resourceType: string): ExportContext;
   jobs?: JobListInstance;
@@ -66,7 +66,7 @@ declare class ExportContext {
    * access, please contact help@twilio.com.
    *
    * @param version - Version of the resource
-   * @param resourceType - The type of communication – Messages, Calls
+   * @param resourceType - The type of communication – Messages, Calls, Conferences, and Participants
    */
   constructor(version: BulkExports, resourceType: string);
 
@@ -95,7 +95,7 @@ declare class ExportInstance extends SerializableClass {
    *
    * @param version - Version of the resource
    * @param payload - The instance payload
-   * @param resourceType - The type of communication – Messages, Calls
+   * @param resourceType - The type of communication – Messages, Calls, Conferences, and Participants
    */
   constructor(version: BulkExports, payload: ExportPayload, resourceType: string);