+++ lib/rest/preview/bulk_exports/export/exportCustomJob.d.ts
@@ -18,7 +18,7 @@ import { SerializableClass } from '../../../../interfaces';
  * access, please contact help@twilio.com.
  *
  * @param version - Version of the resource
- * @param resourceType - The type of communication – Messages, Calls
+ * @param resourceType - The type of communication – Messages, Calls, Conferences, and Participants
  */
 declare function ExportCustomJobList(version: BulkExports, resourceType: string): ExportCustomJobListInstance;
 
@@ -238,7 +238,7 @@ declare class ExportCustomJobInstance extends SerializableClass {
    *
    * @param version - Version of the resource
    * @param payload - The instance payload
-   * @param resourceType - The type of communication – Messages, Calls
+   * @param resourceType - The type of communication – Messages, Calls, Conferences, and Participants
    */
   constructor(version: BulkExports, payload: ExportCustomJobPayload, resourceType: string);