+++ lib/rest/sync/v1/service/syncStream.d.ts
@@ -12,8 +12,6 @@ import { SerializableClass } from '../../../../interfaces';
 import { StreamMessageList } from './syncStream/streamMessage';
 import { StreamMessageListInstance } from './syncStream/streamMessage';
 
-type SyncStreamHideExpiredType = 'true'|'false';
-
 /**
  * Initialize the SyncStreamList
  *
@@ -178,7 +176,6 @@ interface SyncStreamListInstanceCreateOptions {
  *                         Function to process each record. If this and a positional
  *                         callback are passed, this one will be used
  * @property done - Function to be called upon completion of streaming
- * @property hideExpired - Hide expired Sync Streams and show only active ones.
  * @property limit -
  *                         Upper limit for the number of records to return.
  *                         each() guarantees never to return more than limit.
@@ -193,7 +190,6 @@ interface SyncStreamListInstanceCreateOptions {
 interface SyncStreamListInstanceEachOptions {
   callback?: (item: SyncStreamInstance, done: (err?: Error) => void) => void;
   done?: Function;
-  hideExpired?: SyncStreamHideExpiredType;
   limit?: number;
   pageSize?: number;
 }
@@ -201,7 +197,6 @@ interface SyncStreamListInstanceEachOptions {
 /**
  * Options to pass to list
  *
- * @property hideExpired - Hide expired Sync Streams and show only active ones.
  * @property limit -
  *                         Upper limit for the number of records to return.
  *                         list() guarantees never to return more than limit.
@@ -214,7 +209,6 @@ interface SyncStreamListInstanceEachOptions {
  *                         efficient page size, i.e. min(limit, 1000)
  */
 interface SyncStreamListInstanceOptions {
-  hideExpired?: SyncStreamHideExpiredType;
   limit?: number;
   pageSize?: number;
 }
@@ -222,13 +216,11 @@ interface SyncStreamListInstanceOptions {
 /**
  * Options to pass to page
  *
- * @property hideExpired - Hide expired Sync Streams and show only active ones.
  * @property pageNumber - Page Number, this value is simply for client state
  * @property pageSize - Number of records to return, defaults to 50
  * @property pageToken - PageToken provided by the API
  */
 interface SyncStreamListInstancePageOptions {
-  hideExpired?: SyncStreamHideExpiredType;
   pageNumber?: number;
   pageSize?: number;
   pageToken?: string;
@@ -387,4 +379,4 @@ declare class SyncStreamPage extends Page<V1, SyncStreamPayload, SyncStreamResou
   toJSON(): any;
 }
 
-export { SyncStreamContext, SyncStreamHideExpiredType, SyncStreamInstance, SyncStreamInstanceUpdateOptions, SyncStreamList, SyncStreamListInstance, SyncStreamListInstanceCreateOptions, SyncStreamListInstanceEachOptions, SyncStreamListInstanceOptions, SyncStreamListInstancePageOptions, SyncStreamPage, SyncStreamPayload, SyncStreamResource, SyncStreamSolution }
+export { SyncStreamContext, SyncStreamInstance, SyncStreamInstanceUpdateOptions, SyncStreamList, SyncStreamListInstance, SyncStreamListInstanceCreateOptions, SyncStreamListInstanceEachOptions, SyncStreamListInstanceOptions, SyncStreamListInstancePageOptions, SyncStreamPage, SyncStreamPayload, SyncStreamResource, SyncStreamSolution }