+++ CHANGES.md
@@ -1,6 +1,24 @@
 twilio-node changelog
 =====================
 
+[2021-01-13] Version 3.55.0
+---------------------------
+**Api**
+- Add 'Electric Imp v1 Usage' to usage categories
+
+**Conversations**
+- Changed `last_read_message_index` type in Participant's resource **(breaking change)**
+
+**Insights**
+- Added `created_time` to call summary.
+
+**Sync**
+- Remove HideExpired query parameter for filtering Sync Documents with expired **(breaking change)**
+
+**Video**
+- [Rooms] Expose maxConcurrentPublishedTracks property in Room resource
+
+
 [2021-01-06] Version 3.54.2
 ---------------------------
 **Library - Fix**